Fix a security issue regarding dissolutions and a DEX DoS
Dissolutions could be issued specifying an arbitrary purchaser, 
intending to allow agents to propose dissolutions on behalf of others. 
This enables anyone with an active approval, either for the DEX or 
another dissolution, to have a dissolution created for them which would 
drain their wallet. Now, the purchaser must be the proposer.

The DEX now checks before buy orders are placed that the purchaser can 
receive tokens to prevent deadlocks at specific price levels when it 
attempts to execute those trades.