Use Solidity 0.8.13
Solidity 0.8.4 has the new error feature. Solidity 0.8.9 has bug fixes 
making it the minimal viable version. Solidity 0.8.10 improves the 
SMTChecker. Solidity 0.8.11 *introduces abi.encodeCall*, the main 
interesting thing offered in the new versions. Solidity 0.8.13 has a bug 
fix for it

While it is extremely new, it will be of an acceptable age post-audit 
and post-testnet