Remove the requirement for the Frabric and Thread proposal info to be the same
The ABI encoding call was invalid. Since Solidity ABI encoding is offset 
based, it could be made valid, yet only if it was the sole piece of 
dynamic data in the arguments. The next option isn't to manually encode 
it into the data, yet require the tail bytes line up. As long as info 
string is the last argument, this will happen by default, and comply 
with strict encode/decode, and be on-chain verifiable. The problem is 
Solidity SCs don't require strict decode and you can append the info 
string to the data AFTER the info string the Thread will actually use 
without issue. This means you either need to fully parse the relevant 
data or at least know its base length, at which point this is not worth 
it and should instead be relegated to the UI to handle (displaying both 
proposal texts as distinct and part of the actual proposal).

Places info as the last argument, not only due to natural ordering 
expectations, yet also to make manually reading the ABI encoded 
arguments easier (places the data actually executed on first).