Fix fee on transfer support in the DEX
Includes the move from require to Solidity 0.8.4 errors for the DEX 
contracts, as that was the work in progress when these bugs were 
discovered.

The DEX tracked the amount it received, yet then claimed the (presumably 
USD) token amount was the Thread token amount instead of dividing it 
against the price to get the actual order size. In order to ensure users 
are aware of the fees they'll face when transferring, amount has been 
split into payment and minimumAmount, where payment / price will equal 
minimumAmount for non-fee-on-transfer tokens.

Updates comments on ERC20 token behavior and re-entrancy as well.