Efficiently pack data
Moves time fields to uint64, except for auction length which is uint32 
(signifying a max auction length of ~68 years).

Tweaks proposal type to uint16 and changes the commonProposalBit from 
most significant to least significant in the more significant byte. If 
used as a bit flag, there's 8 possible enum types which should be more 
than plenty (right now we have 2 and I don't predict we'll need a 
third). If used as an independent uint8, there's 255 possible types. By 
using the least significant bit this time, we can decide in the future 
if it's a bit flag or an int.