Revamp Auction for security and efficiency
Removes listTransferred and the according balance code for just list 
(increases gas as now it's calling out to transfer for each batch). 
Moves the batch code from FrabricERC20 to Auction to reduce gas (as now 
it's only calling out once).

The previous pattern, which was SLOAD + STATICCALL, would still be 
cheaper than this (STATICCALL CALL STATICCALL) if both were performed in 
aggregate (which is possible). This just reduces complexity and 
therefore attack surface.

Also finally adds ReentrancyGuard to it.