Deduplicate code and fix event checks
Waffle apparently doesn't support multi-events and will frequently 
silently error when the wrong syntax is used, while then forcing use to 
use an archaic syntax that still doesn't work as it's supposed to. It's 
hell.

Renames ProposingParticipantRemoval error to Irremovable, as the first 
isn't actually an issue. The second is a direct reply to the action.

Corrects a bug in DistributionERC20.

Bans fee on transfer for dissolutions/distributions.